Move window insertion logic into backend
