+++ bspwc/xdg_surface_v6.c
@@ -38,20 +38,19 @@ void handle_xdg_shell_v6_surface(struct wl_listener* listener, void* data)
         return;
     }
 
-    struct backend* backend = wl_container_of(listener, backend, new_xdg_shell_v6);
-	struct server* server = backend->server;
-	struct input* input = server->input;
+	char* title = wlr_xdg_surface_v6->toplevel->title;
 
-    wlr_log(
-			L_INFO,
-			"New xdg_v6 toplevel: title=%s, app_id=%s",
-			wlr_xdg_surface_v6->toplevel->title,
-			wlr_xdg_surface_v6->toplevel->app_id
-        );
+    wlr_log(L_INFO, "New xdg_v6 toplevel: title=%s, app_id=%s", title, wlr_xdg_surface_v6->toplevel->app_id);
     wlr_xdg_surface_v6_ping(wlr_xdg_surface_v6);
 
     struct xdg_surface_v6* surface = calloc(1, sizeof(struct xdg_surface_v6));
+	if (surface == NULL)
+	{
+		wlr_log(L_ERROR, "Failed to create surface for %s", title);
+		return;
+	}
 
+	// Wire up handlers and signals
     surface->surface_commit.notify = handle_surface_commit;
     wl_signal_add(&wlr_xdg_surface_v6->surface->events.commit, &surface->surface_commit);
 
@@ -83,44 +82,29 @@ void handle_xdg_shell_v6_surface(struct wl_listener* listener, void* data)
     struct window* window = create_window();
     if (window == NULL)
     {
-		wlr_log(L_ERROR, "Failed to create window for %s", wlr_xdg_surface_v6->toplevel->title);
+		wlr_log(L_ERROR, "Failed to create window for %s", title);
         free(surface);
         return;
     }
 
+	// Fill window with xdg_v6 related stuff
     window->type = XDG_SHELL_V6;
     window->wlr_xdg_surface_v6 = wlr_xdg_surface_v6;
     window->xdg_surface_v6 = surface;
 
-	// Insert on active desktop
-	struct wlr_output* wlr_output = wlr_output_layout_output_at(
-			server->output_layout,
-			input->cursor->wlr_cursor->x,
-			input->cursor->wlr_cursor->y
-		);
-
-	if (wlr_output == NULL)
+	if (title != NULL)
 	{
-		return;
+		window->title = malloc(strlen(wlr_xdg_surface_v6->toplevel->title) + 1);
+		strcpy(window->title, wlr_xdg_surface_v6->toplevel->title);
+	}
+	else
+	{
+		window->title = NULL;
 	}
 
-	// Find the right output
-	struct output* output = NULL;
-	wl_list_for_each(output, &backend->outputs, link)
-	{
-		if (output->wlr_output == wlr_output)
-		{
-			break;
-		}
-	}
-
-	if (output == NULL)
-	{
-		wlr_log(L_ERROR, "Failed to found output to insert window");
-		return;
-	}
-
-	if (insert_window(output, window) == false)
+	// Insert it into the right output
+	struct backend* backend = wl_container_of(listener, backend, new_xdg_shell_v6);
+	if (insert_window(backend, window) == false)
 	{
 		wlr_log(L_ERROR, "Failed to insert window into desktop");
 	}
