+++ include/bspwc/output.h
@@ -34,6 +34,4 @@ void output_destroy_notify(struct wl_listener* listener, void* data);
 void output_frame_notify(struct wl_listener* listener, void* data);
 void new_output_notify(struct wl_listener* listener, void* data);
 
-bool insert_window(struct output* output, struct window* window);
-
 #endif // OUTPUT_H
